# PERF: numexpr doesn't support floordiv, so don't try 