@jreback 2021-04-01
reverted https://github.com/pandas-dev/pandas/pull/40742